relay: Move fakes into relaytest package (#443)
The production tchannel-go package depends on the relay package, which
has a dependency on the testing package. We do not want production
binaries to depend on testing (this also brings in testing flags into
prod binaries), so move all fakes into a separate relaytest package.

This is a pure refactoring change with no logic changes.
