+++ testutils/test_server.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/uber/tchannel-go"
 	"github.com/uber/tchannel-go/raw"
-	"github.com/uber/tchannel-go/relay"
+	"github.com/uber/tchannel-go/relay/relaytest"
 	"github.com/uber/tchannel-go/testutils/goroutines"
 
 	"github.com/davecgh/go-spew/spew"
@@ -55,7 +55,7 @@ type TestServer struct {
 	relayHosts *SimpleRelayHosts
 
 	// relayStats is the channel's relaying stats (if any).
-	relayStats *relay.MockStats
+	relayStats *relaytest.MockStats
 
 	// channels is the list of channels created for this TestServer. The first
 	// element is always the initial server.
@@ -75,7 +75,7 @@ func NewTestServer(t testing.TB, opts *ChannelOpts) *TestServer {
 	ts := &TestServer{
 		TB:            t,
 		channelStates: make(map[*tchannel.Channel]*tchannel.RuntimeState),
-		relayStats:    relay.NewMockStats(),
+		relayStats:    relaytest.NewMockStats(),
 		introspectOpts: &tchannel.IntrospectionOptions{
 			IncludeExchanges:  true,
 			IncludeTombstones: true,
@@ -186,7 +186,7 @@ func (ts *TestServer) CloseAndVerify() {
 
 // AssertRelayStats checks that the relayed call graph matches expectations. If
 // there's no relay, AssertRelayStats is a no-op.
-func (ts *TestServer) AssertRelayStats(expected *relay.MockStats) {
+func (ts *TestServer) AssertRelayStats(expected *relaytest.MockStats) {
 	if !ts.HasRelay() {
 		return
 	}
